+++ app/models/search_builder.rb
@@ -8,6 +8,7 @@ class SearchBuilder < Blacklight::SearchBuilder
 
   # Add queries that excludes everything except for works and collections
   def filter_models(solr_parameters)
+    return if blacklight_params[:all_models] == "true"
     solr_parameters[:fq] ||= []
     solr_parameters[:fq] << "!{!terms f=internal_resource_ssim}#{models_to_solr_clause}"
   end